refactor: js 파일 모듈화
header, sidebar 각각 모듈화 시행
index.js에서 main 함수만 실행
